pyproject.toml

[tool.poetry]
name = "ai-travel-agent"
version = "0.1.1"
description = ""
authors = ["Nir Bar"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "1.0.5"
langgraph = "1.0.2"
grandalf = "^0.8"
sendgrid = "^6.11.0"
streamlit = "^1.38.0"
serpapi = "^0.1.5"
langchain-core = "1.0.4"
langchain-ollama = "1.0.0"
google-genai = "^1.49.0"
langchain-google-genai = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.6.8"
